Revert "fix tabbing"
pushed out of branch, not master - leading to absent revision, found hash collision??

This reverts commit 94b2ad7b85801e37e2df4671cdc4f26b6d7e5506.

+++ clients/oxygen/oxygenclientgroupitemdata.h
@@ -151,7 +151,7 @@ namespace Oxygen
         { return animation().data()->isRunning(); }
 
         //! update button activity
-        void updateButtonActivity( long visibleItem ) const;
+        void updateButtonActivity( int visibleItem ) const;
 
         //! update buttons
         void updateButtons( bool alsoUpdate ) const;
